pyproject.toml
[tool.poetry]
name = "language-wallpaper"
version = "0.1.0"
description = "Updates your desktop backgound daily with a phrase and its translation"
authors = ["unruffled-nightingale <unruffled.nightingale@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
toml = "~0.10"
requests ="~2.27.1"
pillow = "~9.1.1"
bs4 = "~0.0.1"
beautifulsoup4="~4.11.1"
verbecc="~1.7.1"


[tool.poetry.dev-dependencies]
bandit = "*"
black = "*"
coverage = "*"
flake8 = "*"
flake8-pyproject="*"
isort = "*"
mypy = "*"
pre-commit = "*"
pre-commit-hooks = "*"
pytest = "*"
pytest-cov = "*"
pytest-randomly = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
line_length = 120
use_parentheses = "True"

[tool.flake8]
max-line-length = 120
exclude = "venv,.venv,.git,__pycache__,build,dist"

[tool.pytest.ini_options]
addopts = "-ra -v"

[tool.coverage.report]
show_missing = true

[tool.coverage.run]
omit = ["./tests/*"]

[tool.mypy]
disallow_any_unimported = "True"
disallow_any_generics = "True"
disallow_incomplete_defs = "True"
disallow_subclassing_any = "True"
disallow_untyped_calls = "True"
disallow_untyped_defs = "True"
disallow_untyped_decorators = "True"
no_implicit_optional = "True"
pretty = "True"
show_error_codes = "True"
show_error_context = "True"
strict_equality = "True"
strict = "True"
warn_return_any = "True"
warn_unused_ignores = "True"
warn_unreachable = "True"
warn_redundant_casts = "True"
warn_unused_configs = "True"

[[tool.mypy.overrides]]
module = "tests.*"
disallow_any_unimported = "False"
disallow_untyped_calls = "False"
no_implicit_optional = "True"
